requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_login
flask_table
flask_wtf
email_validator
numpy
pandas
scikit-learn
flask_mail
pillow
